pyproject.toml
[project]
name = "ibkr-german-tax-declaration-engine"
version = "3.3.1"
description = "Automate German tax declaration figures from Interactive Brokers Flex Query reports"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "python-dateutil>=2.8.0",
    "requests>=2.25.0",
    "pydantic>=1.8,<2.0",
    "reportlab>=3.5.0",
]

[dependency-groups]
dev = [
    "pymupdf>=1.24.11",
    "pytest",
    "pyyaml",
]
